refactor: 테스트 데이터 생성기 패키지 및 빌드 설정 개선
--- 본문 ---
- 테스트 데이터 생성 관련 클래스들의 패키지명을 `entity`에서 `generator`로 변경하여 역할의 명확성을 높였습니다.
- `LoadTestDataRunner` 클래스명을 `LoadTestDataRunnerTest`로 변경하고 `@Tag("test-data-generator")`를 추가했습니다.
- `build.gradle`에서 `test-data-generator` 태그를 가진 테스트가 기본 `test` 태스크에서 제외되도록 설정하여 GitHub Actions 빌드 시 불필요한 데이터 생성 실행을 방지했습니다.
- `application-test-data.yml`에서 `test-data.generate.enabled` 설정을 제거했습니다.
