+++ src/test/resources/application-test-data.yml
@@ -1,7 +1,3 @@
-test-data:
-  generate:
-    enabled: false
-
 spring:
   datasource:
     url: jdbc:postgresql://localhost:5432/monew?reWriteBatchedInserts=true
